auth

+++ server/routers/snippetRouter.js
@@ -1,10 +1,11 @@
 const express = require("express");
 const router = express.Router();
 const Snippet = require("../models/snippetModel");
+const auth = require("../middleware/auth");
 
-router.get("/", async (req, res) => {
+router.get("/", auth, async (req, res) => {
   try {
-    const snippets = await Snippet.find();
+    const snippets = await Snippet.find({ user: req.user });
 
     res.json(snippets);
   } catch (err) {
@@ -12,10 +13,11 @@ router.get("/", async (req, res) => {
   }
 });
 
-router.post("/", async (req, res) => {
+router.post("/", auth, async (req, res) => {
   try {
     const { title, description, code } = req.body;
 
+    // validation
     if (!description && !code) {
       return res.json({
         errorMessage: "You need to enter at least a some code.",
@@ -26,6 +28,7 @@ router.post("/", async (req, res) => {
       title,
       description,
       code,
+      user: req.user,
     });
 
     const savedSnippet = await newSnippet.save();
@@ -36,7 +39,7 @@ router.post("/", async (req, res) => {
   }
 });
 
-router.put("/:id", async (req, res) => {
+router.put("/:id", auth, async (req, res) => {
   const { title, description, code } = req.body;
   const snippetId = req.params.id;
 
@@ -60,6 +63,10 @@ router.put("/:id", async (req, res) => {
     });
   }
 
+  if (originalSnippet.user.toString() !== req.user) {
+    return res.status(401).json({ errorMessage: "Unauthorized." });
+  }
+
   originalSnippet.title = title;
   originalSnippet.description = description;
   originalSnippet.code = code;
@@ -69,7 +76,7 @@ router.put("/:id", async (req, res) => {
   res.json(savedSnippet);
 });
 
-router.delete("/:id", async (req, res) => {
+router.delete("/:id", auth, async (req, res) => {
   try {
     const snippetId = req.params.id;
 
@@ -85,6 +92,10 @@ router.delete("/:id", async (req, res) => {
       return res.json({
         errorMessage: "No Snippet with this ID was found. ",
       });
+    }
+
+    if (exisitingSnippet.user.toString() !== req.user) {
+      return res.status(401).json({ errorMessage: "Unauthorized" });
     }
 
     await exisitingSnippet.delete();
